Prevent attempted re-enabling of providers breaking existing probes.
It was possible to call enable() on a provider more than once, which would
simply attempt to repeat the work required - creating stubs and the appropriate DOF.
However the kernel will in most cases ignore new DOF, leaving the old provider
in place, while the newly-created stubs are elsewhere.

This change simply no-ops subsequent calls to enable() on a provider, which still
doesn't allow changes to be made, but does prevent inadvertent breaking of existing
probes.