support installing without compiling
Some platforms may not have a compiler, or be appropriate for use with
dtrace-provider. On those environments allow the build to succeed and
rely on the stub implementations instead.

This will squelch all build output, which may be loud and noisey for end
users. If you want to see build output set the environment variable V to
anything.